Pack the Parser project
The Stats project references the Parser. However, Stat's NuGet package
does not include packages which it references, such as the Parser. It
is theoretically possible to include multiple projects in one NuGet
package, but that wasn't working out due to the test projects failing
to reference transitive dependencies (i.e. ones in the Parser). It's
easier to just package and publish a NuGet package for the Parser too.