Reduce amount of indirection when handling events.
Previously, event handlers would append objects to a dictionary and
later the dictionary would be read to aggregate the events into a class.
This was redundant and also terrible since events were stored as
`object` and therefore had to be cast every time they were retrieved.

Remove the events dictionary and instead directly add events to the
processedData instance.